% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paletero-package.R
\docType{package}
\name{paletero-package}
\alias{paletero}
\alias{paletero-package}
\title{paletero: Color palettes and scales.}
\description{
Color palettes and scales.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datasketch/paletero}
  \item Report bugs at \url{https://github.com/datasketch/paletero/issues}
}

}
\author{
\strong{Maintainer}: Juan Pablo Marin Diaz \email{jpmarindiaz@gmail.com}

}
\keyword{internal}
