% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_paleta.R
\name{plot_bars}
\alias{plot_bars}
\title{Create a bar visualization of colors}
\usage{
plot_bars(color_data, label_colors = TRUE)
}
\arguments{
\item{color_data}{Tibble with 'color' and 'distribution' columns}

\item{label_colors}{Whether to add color labels}
}
\value{
A ggplot object
}
\description{
Create a bar visualization of colors
}
\keyword{internal}
