% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_paleta.R
\name{plot_spectrum}
\alias{plot_spectrum}
\title{Create a color spectrum visualization}
\usage{
plot_spectrum(color_data, label_colors = TRUE)
}
\arguments{
\item{color_data}{Tibble with 'color' column}

\item{label_colors}{Whether to add color labels}
}
\value{
A ggplot object
}
\description{
Create a color spectrum visualization
}
\keyword{internal}
