% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview_colors}
\alias{preview_colors}
\title{Color previewing utility}
\usage{
preview_colors(palette, values)
}
\arguments{
\item{values}{A set of values to preview colors for}

\item{pal}{A color mapping function, like those returned from \code{\link{colorNumeric}}, et al}
}
\value{
An HTML-based list of the colors and values
}
\description{
Color previewing utility
}
