% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleta.R
\name{viridis_pal}
\alias{viridis_pal}
\title{#' @export
paleta <- function(palette,  n = NULL, alpha = NULL,
                    reverse = FALSE, colors = NULL){
  if(!palette %in% availablePalettes())
    stop("Palette not available")
  if(palette %in% getBrewerPalettes())
    return(brewer_pal(palette, n = n, alpha = alpha, reverse = reverse))
  if(palette %in% getViridisPalettes())
    return(viridis_pal(palette, n = n, alpha = alpha, reverse = reverse))
  if(palette %in% getPaleteroPalettes())
    return(paletero_pal(palette, n = n, alpha = alpha, reverse = reverse))
  if(palette == "custom")
    return(custom_pal(colors, n = n, alpha = alpha, reverse = reverse))
}}
\usage{
viridis_pal(palette, n = NULL, alpha = NULL, reverse = FALSE)
}
\description{
#' @export
paleta <- function(palette,  n = NULL, alpha = NULL,
                    reverse = FALSE, colors = NULL){
  if(!palette %in% availablePalettes())
    stop("Palette not available")
  if(palette %in% getBrewerPalettes())
    return(brewer_pal(palette, n = n, alpha = alpha, reverse = reverse))
  if(palette %in% getViridisPalettes())
    return(viridis_pal(palette, n = n, alpha = alpha, reverse = reverse))
  if(palette %in% getPaleteroPalettes())
    return(paletero_pal(palette, n = n, alpha = alpha, reverse = reverse))
  if(palette == "custom")
    return(custom_pal(colors, n = n, alpha = alpha, reverse = reverse))
}
}
